Move evals EmbeddedGateway tests to e2e test file (#1398)
This lets us remove the 'e2e_tests' feature from the
dev-dependency, which was forcing this feature on when
running tests at the workspace level
